minus

+++ src/ll_cast.c
@@ -6,7 +6,7 @@
 /*   By: jmeier <jmeier@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/28 10:59:06 by jmeier            #+#    #+#             */
-/*   Updated: 2017/12/12 12:27:46 by jmeier           ###   ########.fr       */
+/*   Updated: 2017/12/12 18:39:28 by jmeier           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,12 +51,12 @@ void	ll_oct_cast(va_list arg, t_all *f)
 	output = (unsigned long long)va_arg(arg, void*);
 	prec_pad = prec_parse(output, f, 8);
 	width_pad = width_parse(output, f, 8);
-	prefix = prefix_parse((int)output, 8, 0, f);
+	prefix = prefix_parse((unsigned int)output, 8, 0, f);
 	if (f->dash == 1)
 	{
 		ft_putstr(prefix);
 		ft_putstr(prec_pad);
-		ft_putoct(output);
+		ft_putoct((unsigned long long)output);
 		ft_putstr(width_pad);
 	}
 	else
@@ -65,7 +65,7 @@ void	ll_oct_cast(va_list arg, t_all *f)
 		ft_putstr(prefix);
 		f->zero == 1 && f->prec_flag == 0 ? ft_putstr(width_pad) : 0;
 		ft_putstr(prec_pad);
-		ft_putoct(output);
+		ft_putoct((unsigned long long)output);
 	}
 	supa_free(prec_pad, width_pad, prefix);
 }
@@ -80,12 +80,12 @@ void	ll_hex_cast(va_list arg, t_all *f)
 	output = (unsigned long long)va_arg(arg, void*);
 	prec_pad = prec_parse(output, f, 16);
 	width_pad = width_parse(output, f, 16);
-	prefix = prefix_parse((int)output, 16, 1, f);
+	prefix = prefix_parse((unsigned int)output, 16, 1, f);
 	if (f->dash == 1)
 	{
 		ft_putstr(prefix);
 		ft_putstr(prec_pad);
-		ft_puthex(output);
+		ft_puthex((unsigned long long)output);
 		ft_putstr(width_pad);
 	}
 	else
@@ -94,7 +94,7 @@ void	ll_hex_cast(va_list arg, t_all *f)
 		ft_putstr(prefix);
 		f->zero == 1 && f->prec_flag == 0 ? ft_putstr(width_pad) : 0;
 		ft_putstr(prec_pad);
-		ft_puthex(output);
+		ft_puthex((unsigned long long)output);
 	}
 	supa_free(prec_pad, width_pad, prefix);
 }
@@ -109,12 +109,12 @@ void	ll_lhex_cast(va_list arg, t_all *f)
 	output = (unsigned long long)va_arg(arg, void*);
 	prec_pad = prec_parse(output, f, 16);
 	width_pad = width_parse(output, f, 16);
-	prefix = prefix_parse((int)output, 16, 0, f);
+	prefix = prefix_parse((unsigned int)output, 16, 0, f);
 	if (f->dash == 1)
 	{
 		ft_putstr(prefix);
 		ft_putstr(prec_pad);
-		ft_putlhex(output);
+		ft_putlhex((unsigned long long)output);
 		ft_putstr(width_pad);
 	}
 	else
@@ -123,7 +123,7 @@ void	ll_lhex_cast(va_list arg, t_all *f)
 		ft_putstr(prefix);
 		f->zero == 1 && f->prec_flag == 0 ? ft_putstr(width_pad) : 0;
 		ft_putstr(prec_pad);
-		ft_putlhex(output);
+		ft_putlhex((unsigned long long)output);
 	}
 	supa_free(prec_pad, width_pad, prefix);
 }
@@ -138,7 +138,7 @@ void	ll_uint_cast(va_list arg, t_all *f)
 	output = (unsigned char)va_arg(arg, void*);
 	prec_pad = prec_parse((unsigned long long)output, f, 10);
 	width_pad = width_parse((unsigned long long)output, f, 10);
-	prefix = prefix_parse((int)output, 10, 0, f);
+	prefix = prefix_parse((unsigned int)output, 10, 0, f);
 	if (f->dash == 1)
 	{
 		ft_putstr(prefix);
